move db tests out of client test module
